texas/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-socketio==5.7.2
